refactor: move experimental rollout engines to canonical rllm/engine/rollout/
The experimental rollout engines (RolloutEngine, ModelOutput, TinkerEngine,
VerlEngine, Completer, TITOCompleter, types) are now the canonical
implementations in rllm/engine/rollout/. The experimental versions were
strict supersets of the non-experimental ones, with additional support for
TITO (token-in-token-out), weight version tracking, and training pipeline
integration.

Key changes:
- Move all engine classes and types from rllm/experimental/rollout/ to
  rllm/engine/rollout/ (the canonical location)
- Add backward-compat shim in rllm/experimental/rollout/__init__.py that
  re-exports from the canonical location with a DeprecationWarning
- Update all 24+ consumer imports from rllm.experimental.rollout to
  rllm.engine.rollout
- Standardize on .is_validation (not .validate) as the attribute name
  across all rollout engine consumers
- OpenAIEngine and FireworksEngine remain in rllm/engine/rollout/ and
  inherit the enhanced base RolloutEngine automatically
