+++ rllm/engine/rollout/__init__.py
@@ -5,7 +5,18 @@ __all__ = [
     "ModelOutput",
     "RolloutEngine",
     "OpenAIEngine",
+    "TinkerEngine",
     "VerlEngine",
+    "Completer",
+    "TITOCompleter",
+    # Token types
+    "TokenInput",
+    "TokenOutput",
+    "TinkerTokenInput",
+    "TinkerTokenOutput",
+    "VerlTokenInput",
+    "VerlTokenOutput",
+    "Tokenizer",
 ]
 
 
@@ -14,6 +25,13 @@ def __getattr__(name):
         from .openai_engine import OpenAIEngine as _OpenAIEngine
 
         return _OpenAIEngine
+    if name == "TinkerEngine":
+        try:
+            from .tinker_engine import TinkerEngine as _TinkerEngine
+
+            return _TinkerEngine
+        except Exception:
+            raise AttributeError(name) from None
     if name == "VerlEngine":
         try:
             from .verl_engine import VerlEngine as _VerlEngine
@@ -21,4 +39,13 @@ def __getattr__(name):
             return _VerlEngine
         except Exception:
             raise AttributeError(name) from None
-    raise AttributeError(name)
+    if name in ("Completer", "TITOCompleter"):
+        from .completer import Completer, TITOCompleter
+
+        return {"Completer": Completer, "TITOCompleter": TITOCompleter}[name]
+    # Token types
+    if name in ("TokenInput", "TokenOutput", "TinkerTokenInput", "TinkerTokenOutput", "VerlTokenInput", "VerlTokenOutput", "Tokenizer"):
+        from . import types as _types
+
+        return getattr(_types, name)
+    raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
