+++ rllm/experimental/engine/sdk_workflow.py
@@ -36,7 +36,7 @@ from rllm.workflows.workflow import TerminationReason, Workflow
 if TYPE_CHECKING:
     from omegaconf import DictConfig
 
-    from rllm.experimental.rollout import RolloutEngine
+    from rllm.engine.rollout import RolloutEngine
 
 logger = logging.getLogger(__name__)
