+++ rllm/experimental/engine/trace_converter.py
@@ -6,7 +6,7 @@ from typing import Any
 from rllm_model_gateway.models import TraceRecord
 
 from rllm.agents.agent import Step, Trajectory
-from rllm.experimental.rollout import ModelOutput
+from rllm.engine.rollout import ModelOutput
 from rllm.tools.tool_base import ToolCall
