+++ rllm/experimental/rollout/__init__.py
@@ -1,17 +1,28 @@
-from typing import TYPE_CHECKING
+"""Backward-compatibility shim — all rollout engines have moved to ``rllm.engine.rollout``.
 
-from .rollout_engine import ModelOutput, RolloutEngine
-from .types import TinkerTokenInput, TinkerTokenOutput, TokenInput, Tokenizer, TokenOutput, VerlTokenInput, VerlTokenOutput
+Importing from ``rllm.experimental.rollout`` still works but will emit a
+``DeprecationWarning``.  Please update your imports to use
+``rllm.engine.rollout`` instead.
+"""
 
-if TYPE_CHECKING:
-    from .tinker_engine import TinkerEngine
-    from .verl_engine import VerlEngine
+import warnings
+
+from rllm.engine.rollout.rollout_engine import ModelOutput, RolloutEngine
+from rllm.engine.rollout.types import TinkerTokenInput, TinkerTokenOutput, TokenInput, Tokenizer, TokenOutput, VerlTokenInput, VerlTokenOutput
+
+warnings.warn(
+    "Importing from 'rllm.experimental.rollout' is deprecated. Use 'rllm.engine.rollout' instead.",
+    DeprecationWarning,
+    stacklevel=2,
+)
 
 __all__ = [
     "ModelOutput",
     "RolloutEngine",
     "TinkerEngine",
     "VerlEngine",
+    "Completer",
+    "TITOCompleter",
     # Token types
     "TokenInput",
     "TokenOutput",
@@ -25,14 +36,21 @@ __all__ = [
 
 def __getattr__(name):
     if name == "TinkerEngine":
-        from .tinker_engine import TinkerEngine as _TinkerEngine
+        try:
+            from rllm.engine.rollout.tinker_engine import TinkerEngine as _TinkerEngine
 
-        return _TinkerEngine
+            return _TinkerEngine
+        except Exception:
+            raise AttributeError(name) from None
     if name == "VerlEngine":
         try:
-            from .verl_engine import VerlEngine as _VerlEngine
+            from rllm.engine.rollout.verl_engine import VerlEngine as _VerlEngine
 
             return _VerlEngine
         except Exception:
             raise AttributeError(name) from None
+    if name in ("Completer", "TITOCompleter"):
+        from rllm.engine.rollout.completer import Completer, TITOCompleter
+
+        return {"Completer": Completer, "TITOCompleter": TITOCompleter}[name]
     raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
