+++ rllm/experimental/verl/verl_backend.py
@@ -33,13 +33,13 @@ from verl.workers.utils.padding import left_right_2_no_padding, no_padding_2_pad
 
 from rllm.agents.agent import Episode
 from rllm.data import Dataset
+from rllm.engine.rollout import RolloutEngine, VerlEngine
 from rllm.experimental.common import (
     AlgorithmConfig,
     collect_reward_and_advantage_from_trajectory_groups,
     simple_timer,
 )
 from rllm.experimental.protocol import BackendProtocol
-from rllm.experimental.rollout import RolloutEngine, VerlEngine
 from rllm.experimental.verl import compute_advantage_verl, transform_episodes_to_dataproto, update_dataproto_with_advantages
 
 if TYPE_CHECKING:
