+++ rllm/trainer/tinker/tinker_backend.py
@@ -23,9 +23,9 @@ from transformers import AutoTokenizer
 
 from rllm.agents.agent import Episode
 from rllm.data import Dataset
+from rllm.engine.rollout import RolloutEngine, TinkerEngine
 from rllm.experimental.common import AlgorithmConfig, simple_timer
 from rllm.experimental.protocol import BackendProtocol
-from rllm.experimental.rollout import RolloutEngine, TinkerEngine
 from rllm.trainer.tinker.tinker_metrics_utils import (
     update_training_metrics,
 )
