+++ tests/integration/test_tinker_gateway_integration.py
@@ -19,7 +19,7 @@ def create_tinker_engine():
     import tinker
     from tinker_cookbook.tokenizer_utils import get_tokenizer
 
-    from rllm.experimental.rollout.tinker_engine import TinkerEngine
+    from rllm.engine.rollout.tinker_engine import TinkerEngine
 
     tokenizer = get_tokenizer(TINKER_MODEL_NAME)
     service_client = tinker.ServiceClient()
